+++ shadow/changelog.md
@@ -1,3 +1,4 @@
 v 0.0.1
 
 - can run sql
+- write table to view data